# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for wallet and transaction management with JWT auth and RBAC.",
        "title": "ExpenseTracker API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
...
        "/wallet": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Shows user's wallet's balance and user id or another user's wallet if the issuer is an admin and inputs an id as a query parameter",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wallet"
                ],
                "summary": "Retrieves user's wallet info",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User's id if admin wants to query another user's wallet",
                        "name": "id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Wallet200Response"
                        }
                    },
                    "400": {
                        "description": "invalid request body",
                        "schema": {
                            "$ref": "#/definitions/dto.Error400"
                        }
                    },
                    "401": {
                        "description": "invalid authorization",
                        "schema": {
                            "$ref": "#/definitions/dto.Error401"
                        }
                    },
                    "403": {
                        "description": "user not authorized to view this wallet",
                        "schema": {
                            "$ref": "#/definitions/dto.Error403"
                        }
                    },
                    "404": {
                        "description": "inputted username not found in database",
                        "schema": {
                            "$ref": "#/definitions/dto.Error404"
                        }
                    },
                    "500": {
                        "description": "failed to correctly query the database or parse claims",
                        "schema": {
                            "$ref": "#/definitions/dto.Error500"
                        }
                    }
                }
            }
        },
        "/wallet/deposit": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Takes inputted amount and adds it to wallet's balance",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wallet"
                ],
                "summary": "Deposit funds to wallet",
                "parameters": [
                    {
                        "description": "Deposit payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.DepositRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Wallet200Response"
                        }
                    },
                    "400": {
                        "description": "invalid request body",
                        "schema": {
                            "$ref": "#/definitions/dto.Error400"
                        }
                    },
                    "401": {
                        "description": "invalid authorization header",
                        "schema": {
                            "$ref": "#/definitions/dto.Error401"
                        }
                    },
                    "500": {
                        "description": "failed to correctly query the database or parse claims",
                        "schema": {
                            "$ref": "#/definitions/dto.Error500"
                        }
                    }
                }
            }
        },
        "/wallet/transfer": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Takes inputted amount from user's wallet and transfers it to another user's wallet",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wallet"
                ],
                "summary": "Transfers funds between users",
                "parameters": [
                    {
                        "description": "Transfer payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.TransferRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Wallet200Response"
                        }
                    },
                    "400": {
                        "description": "invalid request body",
                        "schema": {
                            "$ref": "#/definitions/dto.Error400"
                        }
                    },
                    "401": {
                        "description": "invalid authorization header",
                        "schema": {
                            "$ref": "#/definitions/dto.Error401"
                        }
                    },
                    "404": {
                        "description": "inputted username not found in database",
                        "schema": {
                            "$ref": "#/definitions/dto.Error404"
                        }
                    },
                    "422": {
                        "description": "cannot transfer to self",
                        "schema": {
                            "$ref": "#/definitions/dto.Error422Transfer"
                        }
                    },
                    "500": {
                        "description": "failed to correctly query the database or parse claims",
                        "schema": {
                            "$ref": "#/definitions/dto.Error500"
                        }
                    }
                }
            }
        },
        "/wallet/withdraw": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Takes inputted amount and subtracts it from wallet's balance",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wallet"
                ],
                "summary": "Withdraws funds from wallet",
                "parameters": [
                    {
                        "description": "Withdraw payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.WithdrawRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Wallet200Response"
                        }
                    },
                    "400": {
                        "description": "invalid request body",
                        "schema": {
                            "$ref": "#/definitions/dto.Error400"
                        }
                    },
                    "401": {
                        "description": "invalid authorization header",
                        "schema": {
                            "$ref": "#/definitions/dto.Error401"
                        }
                    },
                    "422": {
                        "description": "not enough balance to withdraw",
                        "schema": {
                            "$ref": "#/definitions/dto.Error422Withdraw"
                        }
                    },
                    "500": {
                        "description": "failed to correctly query the database or parse claims",
                        "schema": {
                            "$ref": "#/definitions/dto.Error500"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.DepositRequest": {
            "type": "object",
            "required": [
                "amount"
            ],
            "properties": {
                "amount": {
                    "type": "integer",
                    "example": 1000
                },
                "category": {
                    "type": "string",
                    "example": "salary"
                },
                "note": {
                    "type": "string",
                    "example": "Don't spend too much!"
                }
            }
        },
        "dto.Error400": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Invalid request body"
                }
            }
        },
        "dto.Error401": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "missing claims"
                }
            }
        },
        "dto.Error403": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "user not authorized to view this wallet"
                }
            }
        },
        "dto.Error404": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "username not found in records"
                }
            }
        },
        "dto.Error422Transfer": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "cannot transfer to self"
                }
            }
        },
        "dto.Error422Withdraw": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "insufficient funds"
                }
            }
        },
        "dto.Error500": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "failed to query database"
                }
            }
        },
        "dto.TransferRequest": {
            "type": "object",
            "required": [
                "amount",
                "to_user"
            ],
            "properties": {
                "amount": {
                    "type": "integer",
                    "example": 10000
                },
                "category": {
                    "type": "string",
                    "example": "Tips"
                },
                "note": {
                    "type": "string",
                    "example": "You earned it!"
                },
                "to_user": {
                    "type": "string",
                    "example": "Adam"
                }
            }
        },
        "dto.Wallet200Response": {
            "type": "object",
            "properties": {
                "balance": {
                    "type": "integer",
                    "example": 20000
                },
                "user_id": {
                    "type": "integer",
                    "example": 3
                }
            }
        },
        "dto.WithdrawRequest": {
            "type": "object",
            "required": [
                "amount"
            ],
            "properties": {
                "amount": {
                    "type": "integer",
                    "example": 500
                },
                "category": {
                    "type": "string",
                    "example": "food"
                },
                "note": {
                    "type": "string",
                    "example": "to buy chicken"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and your JWT.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}